Add random logins who walk randomly
All of the logging in and walking is handled by the WebsocketClient,
who listens to events emitted by the RemoteServerStub.  This demonstrates
that the server controls what happens in the client program through
the WebsocketClient, and that the events used to handle other, pseudo-
remote clients connected to the game server are exactly the same as the
events used to drive the avatar of the client who is playing.